+++ packages/terra-functional-testing/src/terra-cli/wdio/index.js
@@ -47,7 +47,7 @@ const cli = {
         describe: 'A flag to disable the selenium docker service.',
         default: false,
       },
-      ignoreScreenshotMistmatch: {
+      ignoreScreenshotMismatch: {
         type: 'boolean',
         describe: 'A flag to disable failing on screenshot mismatch.',
         default: false,